refactor: orders 상수와 checkout 상수 모듈 분리 (#305)
